Add Header to docs
